refactor(frontend): change location of workout state module
