+++ apps/frontend/app/routes/_dashboard.fitness.exercises.item.$id._index.tsx
@@ -61,7 +61,7 @@ import {
 	getAuthorizationHeader,
 	serverGqlService,
 } from "~/lib/utilities.server";
-import { addExerciseToWorkout, currentWorkoutAtom } from "~/lib/workout";
+import { addExerciseToWorkout, currentWorkoutAtom } from "~/lib/state/workout";
 
 const searchParamsSchema = z.object({
 	defaultTab: z.string().optional(),
